requirements.txt
peewee>=2.10
python-dotenv>=0.7
Sphinx>=1.6